Separate `rpc_info` module in two
Rename `rpc_info` into `rpc_address_file` and move RPC client code used
for checking for other instances to a separate `rpc_uniqueness_check`
module.
